group1/requirements.txt
python-csv
opencv-python
interactive_kit
ipywidgets
matplotlib
numpy
Pillow
scikit-image
imageio
openpyxl
tqdm